NaariKavach_Backend/requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
djoser>=2.3.0
python-socketio>=5.10.0
eventlet>=0.33.3
django-cors-headers>=4.7.0
